+++ lib/src/node/render_error.dart
@@ -10,19 +10,30 @@ import 'utils.dart';
 @anonymous
 class RenderError {
   external String get message;
+  external String get formatted;
   external int get line;
   external int get column;
-  external int get status;
   external String get file;
+  external int get status;
 
   external factory RenderError._(
-      {String message, int line, int column, int status, String file});
+      {String message,
+      String formatted,
+      int line,
+      int column,
+      String file,
+      int status});
 }
 
 RenderError newRenderError(String message,
-    {int line, int column, int status, String file}) {
+    {String formatted, int line, int column, String file, int status}) {
   var error = new RenderError._(
-      message: message, line: line, column: column, status: status, file: file);
+      message: message,
+      formatted: formatted,
+      line: line,
+      column: column,
+      file: file,
+      status: status);
   setToString(error, () => "Error: $message");
   return error;
 }